changes to make the code more modular, calling the operations and classes notebooks from the notebook
