+++ 02-dlt-meta-ingestion/classes.py
@@ -1,10 +1,5 @@
 # Databricks notebook source
-import dlt
-from pyspark.sql.functions import *
-
-# COMMAND ----------
-
-# MAGIC %run ./operations
+# MAGIC %run "./operations"
 
 # COMMAND ----------
 
@@ -12,7 +7,7 @@ class IngestionDLT:
 
     def __init__(
         self
-        ,spark
+        ,spark: SparkSession = spark
         ,env_mode: str = "dev"
         ,catalog: str = "lakehouse"
         ,schema: str = "landing"
@@ -24,12 +19,14 @@ class IngestionDLT:
         self.schema = schema
         self.volume = volume
 
-        use_catalog_schema(catalog = self.catalog, schema = self.schema, env_mode = self.env_mode, verbose = False)
-
-        self.catalog_set = spark.sql("select current_catalog()").collect()[0][0]
+        # use_catalog_schema(catalog = self.catalog, schema = self.schema, env_mode = self.env_mode, verbose = False)
+        if self.env_mode == "prd":
+            self.catalog_set = self.catalog
+        else:
+            self.catalog_set = f"{self.catalog}_{self.env_mode}"
 
     def __repr__(self):
-        return f"""IngestionDLT(env_mode='{self.env_mode}', catalog='{self.catalog}', schema='{self.schema}', volume='{self.volume}')"""
+        return f"""IngestionDLT(env_mode='{self.env_mode}', catalog='{self.catalog_set}', schema='{self.schema}', volume='{self.volume}')"""
 
     def ingest_raw_to_bronze(self, table_name: str, table_comment: str, table_properties: dict, source_folder_path_from_volume: str = "", maxFiles: int = 1000, maxBytes: str = "10g", wholeText: bool = True, options: dict = None):
         """
@@ -52,47 +49,26 @@ class IngestionDLT:
             raw_df = read_stream_raw(spark = spark, path = file_path, maxFiles = maxFiles, maxBytes = maxBytes, wholeText = wholeText, options = options)
 
             bronze_df = (raw_df
-                .withColumn("inputFilename", input_file_name())
+                .withColumn("inputFilename", col("_metadata.file_name"))
+                .withColumn("fullFilePath", col("_metadata.file_path"))
+                .withColumn("fileMetadata", col("_metadata"))
                 .select(
-                    lit(file_path).alias("datasource")
+                    "fullFilePath"
+                    ,lit(file_path).alias("datasource")
                     ,"inputFileName"
                     ,current_timestamp().alias("ingestTime")
                     ,current_timestamp().cast("date").alias("ingestDate")
                     ,"value"
+                    ,"fileMetadata"
                 )
             )
 
             return bronze_df
+        
+    def ingest_raw_to_bronze_synchronous(self, table_names: list, table_comments: list, table_properties: dict, source_folder_path_from_volumes: str, maxFiles: int = 1000, maxBytes: str = "10g", wholeText: bool = True, options: dict = None):
+        """
+            Synchronously ingest from multiple subfolders of the same Volume into more than one bronze table.  Each bronze table created is managed as a streaming Delta Live Table in the same <catalog.schema> as the source volume.  
+        """
+        for i in range(0,len(table_names)):
+            ingest_raw_to_bronze(self = self, table_name = table_names[i], table_comment = table_comments[i], source_folder_path_from_volume = source_folder_path_from_volumes[i], table_properties = table_properties, maxFiles = maxFiles, maxBytes = maxBytes, wholeText = wholeText, options = options)
 
-
-# COMMAND ----------
-
-# pipeline = IngestionDLT(
-#     spark = spark
-#     ,env_mode = "dev"
-#     ,catalog = "lakehouse"
-#     ,schema = "landing"
-#     ,volume = 'dropbox'
-# )
-
-# COMMAND ----------
-
-# pipeline
-
-# COMMAND ----------
-
-# %sql
-# select current_catalog(), current_schema();
-
-# COMMAND ----------
-
-# spark.sql("select current_catalog()").collect()[0][0]
-
-# COMMAND ----------
-
-# pipeline.ingest_raw_to_bronze(
-#     table_name="dropbox_bronze"
-#     ,table_comment="A full text record of every file that has landed in our dropbox folder."
-#     ,table_properties=None
-#     ,source_folder_path_from_volume=""
-# )
